+++ library/flow/min_cost_max_flow.h
@@ -5,7 +5,7 @@ struct min_cost_max_flow {
 	typedef long long ll;
 	struct edge {
 		ll a, b, cap, cost, flow;
-		size_t back;
+		int back;
 	};
 	vector<edge> e;
 	vector<vector<ll>> g;
@@ -15,11 +15,11 @@ struct min_cost_max_flow {
 		g.resize(n);
 	}
 	void add_edge(ll a, ll b, ll cap, ll cost) {
-		edge e1 = {a, b, cap, cost, 0, ssize(g[b]) };
-		edge e2 = {b, a, 0, -cost, 0, ssize(g[a]) };
-		g[a].push_back((ll) ssize(e));
+		edge e1 = {a, b, cap, cost, 0, (int)ssize(g[b]) };
+		edge e2 = {b, a, 0, -cost, 0, (int)ssize(g[a]) };
+		g[a].push_back(ssize(e));
 		e.push_back(e1);
-		g[b].push_back((ll) ssize(e));
+		g[b].push_back(ssize(e));
 		e.push_back(e2);
 	}
 	// returns {flow, cost}
@@ -27,7 +27,7 @@ struct min_cost_max_flow {
 		ll flow = 0, cost = 0;
 		while (flow < k) {
 			vector<ll> id(n, 0), d(n, INF), q(n), p(n);
-			vector<size_t> p_edge(n);
+			vector<int> p_edge(n);
 			ll qh = 0, qt = 0;
 			q[qt++] = s;
 			d[s] = 0;
@@ -35,7 +35,7 @@ struct min_cost_max_flow {
 				ll v = q[qh++];
 				id[v] = 2;
 				if (qh == n) qh = 0;
-				for (size_t i = 0; i < ssize(g[v]); i++) {
+				for (int i = 0; i < ssize(g[v]); i++) {
 					edge& r = e[g[v][i]];
 					if (r.flow < r.cap && d[v] + r.cost < d[r.b]) {
 						d[r.b] = d[v] + r.cost;
@@ -56,12 +56,12 @@ struct min_cost_max_flow {
 			ll addflow = k - flow;
 			for (ll v = t; v != s; v = p[v]) {
 				ll pv = p[v];
-				size_t pr = p_edge[v];
+				int pr = p_edge[v];
 				addflow = min(addflow, e[g[pv][pr]].cap - e[g[pv][pr]].flow);
 			}
 			for (ll v = t; v != s; v = p[v]) {
 				ll pv = p[v];
-				size_t pr = p_edge[v], r = e[g[pv][pr]].back;
+				int pr = p_edge[v], r = e[g[pv][pr]].back;
 				e[g[pv][pr]].flow += addflow;
 				e[g[v][r]].flow -= addflow;
 				cost += e[g[pv][pr]].cost * addflow;
